add windows push configs

+++ push-configs.ps1
@@ -1,10 +1,21 @@
-$WinVimpath = Join-Path -PATH $env:LOCALAPPDATA -ChildPath "\nvim\"
+$RepoPath = "T:\Repository\dotfiles"
 
-$RepoVimpath = "T:\Repository\dotfiles\nvim\"
+$WinVimpath = Join-Path -PATH $env:LOCALAPPDATA -ChildPath "\nvim\"
+$RepoVimpath = Join-Path -PATH $RepoPath -ChildPath "\nvim\"
+
+$WinGlazepath = Join-Path -PATH $env:USERPROFILE -ChildPath "\.glzr\glazewm\"
+$RepoGlazepath = Join-Path -PATH $RepoPath -ChildPath "\glazewm\"
 
 Write-Output "Writing nvim config from $RepoVimpath to $WinVimpath"
 
+if ( Test-Path -Path $WinVimpath ){
+    Remove-Item -Path $WinVimpath -Recurse
+}
+Copy-Item -PATH $RepoVimpath -Destination $WinVimpath -Recurse 
 
-# TODO: 'if valid path $WinPath, then delete everything in it
-Remove-Item -Path $WinVimpath -Recurse
-Copy-Item -PATH $RepoVimpath -Destination $WinVimpath -Recurse -Force
+Write-Output "Writing glazeWM config from $RepoGlazepath to $WinGlazepath"
+
+if ( Test-Path -Path $WinGlazepath) {
+    Remove-Item -Path $WinGlazepath -Recurse
+}
+Copy-Item -PATH $RepoGlazepath -Destination $WinGlazepath -Recurse
